tools/requirements.txt
# Copyright (C) <2023-2023> Intel Corporation
# SPDX-License-Identifier: MIT

fabric
pyyaml
docker
